Refactor sbatch_args handling: move to separate module and update imports
